+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 * Improve the webhook:set command so that it accepts the hostname. The webhook URL will be generated automatically. If url or hostname is not passed then command tries to generate url based on [request context](https://symfony.com/doc/current/routing.html#generating-urls-in-commands)
 * Move config from yaml to php files
 * Allow to set update type for webhook command
+* Allow to set timeout for api instances
 
 5.0.0 (2022-01-21)
 ------------------